% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_arrange.R
\name{category_arrange}
\alias{category_arrange}
\title{A Category Arrange Helper Function}
\arguments{
\item{category}{Vector of categories to be arranged.}
}
\description{
This function takes a vector of categories and rearranges them in alphabetical order
ver. May 26 2021
}
\examples{
category_arrange(category)
}
\keyword{wide}
