% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_function.R
\name{read_function}
\alias{read_function}
\title{A READ Function}
\usage{
read_function(survey, iso_code3, year)
}
\arguments{
\item{survey}{Survey types (e.g. MICS, DHS, etc.) of this dataset}

\item{iso_code3}{Country ISO3 code of this dataset}

\item{year}{Year of this dataset}
}
\description{
This function reads in a batch of microdata files with an input vector of "survey types", "iso3 codes", and "years".
Please make sure to run the function in a working directory where it includes microdata files.
ver. Mar 04 2021
}
\examples{
read_function(DHS, COD, 2013)
}
\keyword{wide}
