% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function}
\alias{test_function}
\title{A TEST Function}
\usage{
test_function(test = TRUE)
}
\arguments{
\item{test}{Is this test? Defaults to TRUE.}
}
\description{
This function allows you to test
}
\examples{
test_function()
}
\keyword{test}
