% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_income_groups.R
\name{update_income_groups}
\alias{update_income_groups}
\title{AN UPDATE INCOME GROUPS Function}
\arguments{
\item{df}{dataframe}
}
\description{
This function gets and updates the world bank income groups
ver. June 10 2021
}
\examples{
update_income_groups(df)
}
\keyword{wide}
