% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_function.R
\docType{data}
\name{val_function}
\alias{val_function}
\alias{df}
\title{A VAL (validation) Function}
\format{
An object of class \code{data.frame} with 95949 rows and 43 columns.
}
\usage{
df
}
\arguments{
\item{var}{Variable name to be validated from this dataset}
}
\description{
This function validates certain conditions of the variable values.
ver. May 11 2021
}
\examples{
val_function(age)
}
\keyword{wide}
