% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_function.R
\name{read_function}
\alias{read_function}
\title{A READ Function}
\arguments{
\item{survey}{Survey types (e.g. MICS, DHS, etc.) of this dataset}

\item{iso_code3}{Country ISO3 code of this dataset}

\item{year}{Year of this dataset}
}
\description{
This function reads in a batch of microdata files with an input vectors of "survey types", "iso3 codes", and "years".
Please make sure to change the path where data files are located.
ver. May 31 2021
}
\examples{
read_function(c("AFG", "BLZ"), c(2010, 2011), c("MICS"), "/Users/sunminlee/Desktop/gemr/gemRpackage/microdata/")
}
\keyword{wide}
